Make resource routes and remove client file
