+++ app/routes/blog.$slug.tsx
@@ -1,21 +1,23 @@
-import { json, type LoaderArgs } from "@remix-run/node";
+import { type LoaderArgs } from "@remix-run/node";
 import { useLoaderData } from "@remix-run/react";
 import { getMDXComponent } from "mdx-bundler/client";
-import { getPost } from "blog/client";
 import { useMemo } from "react";
 import invariant from "tiny-invariant";
 import { publishDateFormatter } from "~/utils";
+import { Mdx } from "blog/parser";
 
 export const loader = async function ({ params, request }: LoaderArgs) {
   const { slug } = params;
   invariant(typeof slug === "string", "Missing slug");
-  const post = await getPost(new URL(request.url).origin, slug);
+  const post = (await fetch(
+    `${new URL(request.url).origin}/resource/get-blog-post/${slug}`
+  ).then((response) => response.json())) as Promise<Mdx>;
   invariant(post !== undefined, "Could not find post");
-  return json({ post });
+  return post;
 };
 
 export default function BlogSlugRoute() {
-  const { post } = useLoaderData<typeof loader>();
+  const post = useLoaderData<typeof loader>();
   const { title, author, date } = post.frontmatter;
   const { code } = post;
   const Component = useMemo(() => getMDXComponent(code), [code]);
